% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppCreditCalc.R
\name{calcIdealWorstCaseValueByPath}
\alias{calcIdealWorstCaseValueByPath}
\title{calcIdealWorstCaseValueByPath}
\usage{
calcIdealWorstCaseValueByPath(lstPaths, periodName = "Jun_17",
  onOrOff = "OFF", ftpRoot = LocalDriveDaPrice, numHours = NULL)
}
\arguments{
\item{ftpRoot}{}
}
\description{
calcIdealWorstCaseValueByPath
}
\examples{
aPath <- list(Source = 'WR.MW.SMOK2.CU', Sink = 'SPRM_SPRM')
dateRange <- c('2016-06-01', '2016-07-01')
calcIdealWorstCaseValueByPath(aPath, dateRange = dateRange)

}
