% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppPrice.R
\name{getDfDaCongestDistribution}
\alias{getDfDaCongestDistribution}
\title{getDistribution
returning the distribution of MCC Sink - MCC Source given the path descriptions specified in lstPaths
it will look at the yearOffset if available,}
\usage{
getDfDaCongestDistribution(lstPaths, periodName = "Jun_17", onOrOff = "OFF",
  yearOffset = 1, ftpRoot = FTPROOT_DAPRICE, quantiles = c(0.1, 0.25,
  0.5))
}
\arguments{
\item{lstPaths}{cannot be blank, must be specified with a list of list where Source and Sink must be specified}

\item{periodName}{e.g. Jun_17}

\item{onOrOff}{e.g. OFF}

\item{yearOffset}{e.g. 1 or 2}
}
\description{
getDistribution
returning the distribution of MCC Sink - MCC Source given the path descriptions specified in lstPaths
it will look at the yearOffset if available,
}
