% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppPrice.R
\name{getDfDaPriceSpp}
\alias{getDfDaPriceSpp}
\title{getDfDaPriceSpp}
\usage{
getDfDaPriceSpp(dateRange, ftpRoot = FTPROOT_DAPRICE)
}
\arguments{
\item{dateRange}{a 2-tuple that specifies a from date and a to date}

\item{ftpRoot}{}
}
\value{
a dataframe that gives the price information on all SPP nodes
}
\description{
getDfDaPriceSpp
}
