% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppCreditCalc.R
\name{getDfRefPriceByPeriod}
\alias{getDfRefPriceByPeriod}
\title{getDfRefPriceByPeriod}
\usage{
getDfRefPriceByPeriod(periodName = "Jun_17", onOrOff = "OFF",
  ftpRoot = FTPROOT_REFPRICE, lstPaths = NULL)
}
\arguments{
\item{periodName}{the periodName referred in the Allocation or Auction file. e.g. 'Jun_17'}

\item{onOrOff}{either ON or OFF}

\item{ftpRoot}{points to the ftp site, allows callers to point to local directory if the files have been downloaded}
}
\value{
a data frame containing the reference price table
}
\description{
getDfRefPriceByPeriod
}
