% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppPeriods.R
\name{getNumHours}
\alias{getNumHours}
\title{getNumHours}
\usage{
getNumHours(periodName = "Jun_17", onOrOff = "OFF")
}
\arguments{
\item{periodName}{e.g. 'Jun_17'}

\item{onOrOff}{e.g. 'OFF'}
}
\description{
getNumHours
}
