% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppPeriods.R
\name{getPeriodInfo}
\alias{getPeriodInfo}
\title{getPeriodInfo
returns a data frame that describes more information give a periodName
for example a periodName of 'Jun_17' will return the following list:
  Period : 'Jun'
  FromDate: (date object) '2017-06-01'
  ToDate: (date object) '2017-07-01'
  PeriodFullName: 'June'}
\usage{
getPeriodInfo(periodName = "Jun_17")
}
\arguments{
\item{periodName}{}
}
\value{
a list decsribing more details about the periodName
}
\description{
getPeriodInfo
returns a data frame that describes more information give a periodName
for example a periodName of 'Jun_17' will return the following list:
  Period : 'Jun'
  FromDate: (date object) '2017-06-01'
  ToDate: (date object) '2017-07-01'
  PeriodFullName: 'June'
}
\examples{
getPeriodInfo('Jun_17')
getPeriodInfo('Fall_17')
getPeriodInfo('Spring_18')
}
