% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtoCal.R
\name{getRtoCalendar}
\alias{getRtoCalendar}
\title{create a master calendar for a specific RTO}
\usage{
getRtoCalendar(marketName, fromDate, toDate = NULL,
  props = STANDARD_CAL_PROPERTIES, byMinutes = NULL)
}
\arguments{
\item{marketName}{an RTO name: MISO, SPP, CAISO,}

\item{fromDate}{a date}

\item{toDate}{a date, if NULL then one-day calendar is assumed.}

\item{props}{a character vector that lists the requested properties}
}
\description{
\code{getRtoCalendar} creates a master calendar and allow users to specify the calendar properties
}
\details{
Recognized properties are:
START_DT, START_DT.GMT, DATE, DATE.YEARMONTH, END_DT, HOURENDING, TIMEOFUSE, SEASON, PO_INTERVAL, PLANNING_YEAR
}
\examples{
cal <- getRtoCalendar('SPP', '2017-01-01','2017-04-01', props=c('DATE.YEARMONTH', 'START_DT', 'START_DT.GMT', 'TIMEOFUSE'))

}
